Update ResizableArrayBag.java
The difference bag only works rn if the data points are on the same node and not in the whole thing. I think is a good launch point for the code at least before we go and fix the code and stuff. But I figure it's a while loop and while the length of the first bag is not done it runs through each element and if it's equal to any removes itself from the equation.